Source Code and SWC Update ( T:20110628-19 )
Added a mini profiler that will appear as a 40px high bar at the top of the SWF window.
To enabled the profiler set the config.profile value to TRUE when configuring the game.
Example: config.profile=true

The profiler provides the following information:
Frame Rate: The current and average FPS of the game.
Frame Time: The current and average amount of time (in milliseconds) it took to update and render the game.
Memory: The current and peak amount of RAM used by the game.
Rendering Mode: Whether the game is being rendered with software or hardware (GPU).

Added an option to change the background color of the screens.
To set the background color, set the config.gameBackground value to a RGB value
when configuring the game.
Example: config.gameBackground = 0xFFAA00;

Classes Modified ...

+ GameApp
+ GameConfig
+ GameGraphics
+ GameProfiler

Misc Files Added ...

+ framework/assets/profiler/graphics.png